requirements.txt
absl-py==1.4.0
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
command-not-found==0.3
cryptography==3.4.8
cvxopt==1.3.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
Flask==2.3.2
flatbuffers==23.5.9
fonttools==4.29.1
fs==2.4.12
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.55.0
h5py==3.8.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
imageio==2.28.1
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.1.2
jax==0.4.10
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
libclang==16.0.0
lxml==4.8.0
lz4==3.1.3+dfsg
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
ml-dtypes==0.1.0
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.1
notebook==6.4.8
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==8.6.0.163
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python-headless==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.19.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-build==0.17.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.8.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.13.1
testpath==0.5.0
tf-explain==0.3.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.6.0
tzdata==2023.3
ubuntu-advantage-tools==27.11.2
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.16
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.4
widgetsnbextension==2.0.0
wrapt==1.14.1
zipp==1.0.0
